fibonacci 3

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,15 +8,15 @@
  */
 int main(void)
 {
-	unsigned long num1 = 0, num2 = 1;
-	unsigned long next;
+	unsigned long long num1 = 0, num2 = 1;
+	unsigned long long next;
 	int i;
 
 	for (i = 0; i < 98; ++i)
 	{
 		next = num1 + num2;
 
-		printf("%ld, ", next);
+		printf("%lld, ", next);
 		num1 = num2;
 		num2 = next;
 	}
